requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1605055780383/work
certifi==2020.12.5
chardet==4.0.0
Django @ file:///tmp/build/80754af9/django_1606860386887/work
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.12.2
idna==2.10
Pillow==8.0.1
psycopg2 @ file:///C:/ci/psycopg2_1608147654043/work
pytz @ file:///tmp/build/80754af9/pytz_1606604771399/work
requests==2.25.1
six==1.15.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
urllib3==1.26.2
wincertstore==0.2
